поправил чтобы удоства тоже менялись
